ticket raise 9 Jan 2023

+++ src/main/java/com/authentication/model/Ticket.java
@@ -10,9 +10,10 @@ import java.util.Date;
 @Entity
 @Data
 public class Ticket {
+
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
 
     @CreatedDate
     @CreationTimestamp
